+++ infra/modules/providers/azure/aks/main.tf
@@ -37,7 +37,7 @@ resource "random_id" "main" {
 resource "azurerm_log_analytics_workspace" "main" {
   count = var.log_analytics_id == "" ? 1 : 0
 
-  name                = lower("${var.name}")
+  name                = lower(var.name)
   resource_group_name = data.azurerm_resource_group.main.name
   location            = data.azurerm_resource_group.main.location
   sku                 = "PerGB2018"
@@ -81,7 +81,7 @@ resource "azurerm_kubernetes_cluster" "main" {
     name                = "default"
     node_count          = var.agent_vm_count
     vm_size             = var.agent_vm_size
-    os_disk_size_gb     = 30
+    os_disk_size_gb     = var.agent_vm_disk
     vnet_subnet_id      = var.vnet_subnet_id
     enable_auto_scaling = var.auto_scaling_default_node
     max_pods            = var.max_pods