+++ infra/templates/osdu-r3-mvp/data_partition/main.tf
@@ -33,9 +33,29 @@
 // *** WARNING  ****
 
 terraform {
-  required_version = ">= 0.12"
+  required_version = ">= 0.14"
+
   backend "azurerm" {
     key = "terraform.tfstate"
+  }
+
+  required_providers {
+    azurerm = {
+      source  = "hashicorp/azurerm"
+      version = "=2.41.0"
+    }
+    azuread = {
+      source  = "hashicorp/azuread"
+      version = "=1.1.1"
+    }
+    random = {
+      source  = "hashicorp/random"
+      version = "=2.3.1"
+    }
+    null = {
+      source  = "hashicorp/null"
+      version = "=3.0.0"
+    }
   }
 }
 
@@ -43,26 +63,8 @@ terraform {
 # Providers
 #-------------------------------
 provider "azurerm" {
-  version = "=2.29.0"
   features {}
 }
-
-provider "azuread" {
-  version = "=1.0.0"
-}
-
-provider "random" {
-  version = "~>2.2"
-}
-
-provider "external" {
-  version = "~> 1.0"
-}
-
-provider "null" {
-  version = "~>2.1.0"
-}
-
 
 
 #-------------------------------
@@ -86,10 +88,11 @@ locals {
   resource_group_name = format("%s-%s-%s-rg", var.prefix, local.workspace, random_string.workspace_scope.result)
   retention_policy    = var.log_retention_days == 0 ? false : true
 
-  storage_name      = "${replace(local.base_name_21, "-", "")}data"
-  sdms_storage_name = "${replace(local.base_name_21, "-", "")}sdms"
-  cosmosdb_name     = "${local.base_name}-db"
-  sb_namespace      = "${local.base_name_21}-bus"
+  storage_name        = "${replace(local.base_name_21, "-", "")}data"
+  sdms_storage_name   = "${replace(local.base_name_21, "-", "")}sdms"
+  ingest_storage_name = "${replace(local.base_name_21, "-", "")}ingest"
+  cosmosdb_name       = "${local.base_name}-db"
+  sb_namespace        = "${local.base_name_21}-bus"
 
   eg_sbtopic_subscriber   = "servicebusrecordstopic"
   eventgrid_name          = "${local.base_name_21}-grid"
@@ -100,7 +103,6 @@ locals {
     data.terraform_remote_state.central_resources.outputs.principal_objectId
   ]
 }
-
 
 
 #-------------------------------
@@ -132,7 +134,6 @@ resource "random_string" "workspace_scope" {
 }
 
 
-
 #-------------------------------
 # Resource Group
 #-------------------------------
@@ -143,7 +144,6 @@ resource "azurerm_resource_group" "main" {
   tags = var.resource_tags
   lifecycle { ignore_changes = [tags] }
 }
-
 
 
 #-------------------------------
@@ -158,7 +158,8 @@ module "storage_account" {
   kind                = "StorageV2"
   replication_type    = var.storage_replication_type
 
-  resource_tags = var.resource_tags
+  resource_tags  = var.resource_tags
+  blob_cors_rule = var.blob_cors_rule
 }
 
 // Add Access Control to Principal
@@ -211,6 +212,38 @@ resource "azurerm_role_assignment" "sdms_storage_data_contributor" {
   scope                = module.sdms_storage_account.id
 }
 
+module "ingest_storage_account" {
+  source = "../../../modules/providers/azure/storage-account"
+
+  name                = local.ingest_storage_name
+  resource_group_name = azurerm_resource_group.main.name
+  container_names     = []
+  kind                = "StorageV2"
+  replication_type    = var.storage_replication_type
+
+  resource_tags = var.resource_tags
+}
+
+// Add Access Control to Principal
+resource "azurerm_role_assignment" "ingest_storage_access" {
+  count = length(local.rbac_principals)
+
+  role_definition_name = "Contributor"
+  principal_id         = local.rbac_principals[count.index]
+  scope                = module.ingest_storage_account.id
+}
+
+// Add Data Contributor Role to Principal
+resource "azurerm_role_assignment" "ingest_storage_data_contributor" {
+  count      = length(local.rbac_principals)
+  depends_on = [azurerm_role_assignment.ingest_storage_access]
+
+  role_definition_name = "Storage Blob Data Contributor"
+  principal_id         = local.rbac_principals[count.index]
+  scope                = module.ingest_storage_account.id
+}
+
+
 #-------------------------------
 # CosmosDB
 #-------------------------------
@@ -238,7 +271,6 @@ resource "azurerm_role_assignment" "cosmos_access" {
 }
 
 
-
 #-------------------------------
 # Azure Service Bus
 #-------------------------------
@@ -252,7 +284,6 @@ module "service_bus" {
 
   resource_tags = var.resource_tags
 }
-
 
 // Add Access Control to Principal
 resource "azurerm_role_assignment" "sb_access" {
@@ -300,6 +331,7 @@ resource "azurerm_eventgrid_event_subscription" "service_bus_topic_subscriber" {
 
   service_bus_topic_endpoint_id = lookup(module.service_bus.topicsmap, "recordstopiceg")
 }
+
 
 #-------------------------------
 # Locks